+++ R/funs-conversion.R
@@ -1,23 +1,23 @@
 pl_as.numeric <- function(x, ...) {
   check_empty_dots(...)
-  x$cast(polars::pl$Float64, strict = FALSE)
+  x$cast(pl$Float64, strict = FALSE)
 }
 
 # This one has strict = TRUE because R converts "1.5" (character) to 1 (int)
 # while polars returns null.
 pl_as.integer <- function(x, ...) {
   check_empty_dots(...)
-  x$cast(polars::pl$Int32, strict = TRUE)
+  x$cast(pl$Int32, strict = TRUE)
 }
 
 pl_as.logical <- function(x, ...) {
   check_empty_dots(...)
-  x$cast(polars::pl$Boolean, strict = FALSE)
+  x$cast(pl$Boolean, strict = FALSE)
 }
 
 pl_as.character <- function(x, ...) {
   check_empty_dots(...)
-  x$cast(polars::pl$String, strict = FALSE)
+  x$cast(pl$String, strict = FALSE)
 }
 
 pl_as.Date <- function(x, format, ...) {