+++ R/global_options.R
@@ -19,6 +19,12 @@
 #'   - the session may crash or experience a severe slowdown when the data is
 #'     converted to R (especially if the input is a LazyFrame).
 #'
+#' * `tidypolars_record_query` controls whether the polars query performed in
+#'   the background is recorded so that it can be displayed with
+#'   [show_query()]. The default is `TRUE`. Setting this option to `FALSE`
+#'   removes the (small) overhead of recording the query, but [show_query()]
+#'   will then error on objects created while the option was `FALSE`.
+#'
 #' The package `polars` also contains several global options that may be useful,
 #' such as changing the default behavior when converting Int64 values to R:
 #' <https://pola-rs.github.io/r-polars/man/polars_options.html>.